chore: generate baged
