+++ package.json
@@ -79,11 +79,5 @@
   },
   "publishConfig": {
     "access": "public"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "pnpm test:watch",
-      "pre-push": "pnpm test:coverage"
-    }
   }
 }
